refactor: reorganize imports and consolidate enumerations
