+++ app/GX_integration/invocations/invocation_service.py
@@ -10,15 +10,22 @@ from starlette.status import HTTP_204_NO_CONTENT
 
 from app.galaxy import GalaxyClient
 from app.persistence import MongoStore
-from app.api.enums import CollectionNames
 from app.GX_integration.workflows.workflow_manager import WorkflowManager
 from app.api.schemas import invocation, workflow
 from app.orchestration.invocation_cache import InvocationCache
 from app.orchestration.invocation_tasks import InvocationBackgroundTasks
-from app.enumerations import NumericLimits, InvocationStates
+
+from app.enumerations import (
+    NumericLimits,
+    InvocationStates,
+    CollectionNames
+    )
 from app.api.socket_manager import SocketManager
 
-from app.exceptions import InternalServerErrorException, NotFoundException
+from app.exceptions import (
+    InternalServerErrorException,
+    NotFoundException
+    )
 
 from app.GX_integration.invocations.data_manager import InvocationDataManager
 from app.GX_integration.invocations.tracker import InvocationTracker
