+++ app/GX_integration/workflows/workflow_invocation.py
@@ -13,12 +13,25 @@ load_dotenv()
 from sys import path
 path.append('.')
 
-from bioblend.galaxy.objects.wrappers import Workflow, Invocation, History, Dataset, DatasetCollection
+from bioblend.galaxy.objects.wrappers import (
+    Workflow,
+    Invocation,
+    History,
+    Dataset,
+    DatasetCollection
+    )
+
+from app.enumerations import (
+    NumericLimits,
+    InvocationTracking,
+    JobState,
+    InvocationStates,
+    SocketMessageEvent,
+    SocketMessageType
+    )
 
 from app.galaxy import GalaxyClient
 from app.api.socket_manager import SocketManager
-from app.api.enums import SocketMessageEvent, SocketMessageType
-from app.enumerations import NumericLimits, InvocationTracking, JobState, InvocationStates
 
 class WorkflowInvocationHandler:
     """Handles workflow invocation execution, monitoring, and result collection"""
