+++ app/main.py
@@ -18,16 +18,28 @@ from starlette.status import (
     HTTP_502_BAD_GATEWAY
     )
 
-from fastapi import FastAPI, Request, HTTPException, Query, WebSocket, WebSocketDisconnect, Response
+from fastapi import (
+    FastAPI,
+    Request,
+    HTTPException,
+    Query,
+    WebSocket,
+    WebSocketDisconnect,
+    Response
+    )
 from fastapi.openapi.utils import get_openapi
 from app.mcp_client.chatbot import ChatSession, initialize_session
 
 from app.utils import import_published_workflows
 from app.log_setup import configure_logging
-from app.api.middleware import JWTGalaxyKeyMiddleware, RateLimiterMiddleware, DomainCORSMiddleware
+from app.api.middleware import (
+    JWTGalaxyKeyMiddleware,
+    RateLimiterMiddleware,
+    DomainCORSMiddleware
+    )
 from app.api.api import api_router 
 from app.api.socket_manager import ws_manager
-from app.api.enums import SocketMessageEvent
+from app.enumerations import SocketMessageEvent
 from app.orchestration.invocation_cache import InvocationCache
 from app.orchestration.invocation_tasks import InvocationBackgroundTasks
